requirements.txt
streamlit
numpy
scipy
soundfile
librosa
pydub
openai-whisper
torch